feat: Sprites map and shapes map
Put Sprites/Shape in a map<name, sprite/shape>

	modified:   src/Graphic/IDisplayModule.hpp
	modified:   src/Graphic/SFML/SfmlModule.cpp
	modified:   src/Graphic/SFML/SfmlModule.hpp

+++ src/Graphic/IDisplayModule.hpp
@@ -28,7 +28,7 @@ public:
     virtual ~IDisplayModule() = default;
 
     virtual std::string getName() const = 0;
-    virtual void draw(Drawable drawable, std::pair<int, int> position, int size) = 0;
+    virtual void draw(Drawable drawable, std::pair<float, float> position, int size) = 0;
     virtual void init() = 0;
     virtual void destroy() = 0;
     virtual events_e pollEvent() = 0;
